pyproject.toml
[tool.poetry]
name = "api-fh"
version = "0.1.0"
description = ""
authors = ["Artem Safonov <artyomkaro4e@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.4"}
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
aiomysql = "^0.2.0"
pydantic-settings = "^2.6.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
sqladmin = "^0.20.1"
itsdangerous = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
